Adds better social network link icon support
Updates config.yaml to use a new format, where the owner name and social
link data are in a contiguous object.
Moved social link icon html back into a partial since the code is
repetitive and needs variables. The partialed version takes up less
space and is cleaner, due to the lack of repetition in the arguments.
Adds support for changing social link icon size from the site params.