Improves project section rendering
Creates a new partial, card-project.html, that handles how projects are
displayed in section views.
Creates a section page for projects in the base site hierarchy that
renders projects in a two-card row regardless of screen size.
Updates the home page's recent projects section to render projects in
the same two-card row that is used in the project page.
Adds new parameters to super-slash-n-grab.md for use in the section
pages.
Adjusts content flow in super-slash-n-grab.md in order to better support
section summaries (if I decide to use them later).