Adds Akhet project
Adds akhet.md, which contains the content for the Akhet project.
Adds the only screenshot from the old site for Akhet to
static\images\akhet\in-game.jpg.
Adds a new folder to the project root, static_raw, for raw versions of
static assets.
Adds Photoshop file for the Akhet logo to static_raw.
Adds .png version of Akhet logo to static/images.