Adds nav buttons to bottom of post pages
Adds id tag set to the post's file name to the div of each post summary
card in card-post-summary.html.
Adds three buttons to the bottom of single.html. One pointing to the
previous post, one to the next post, and one that goes back to the post
list page with an anchor to the relevant post.