Adds markdown flowchart support
Adds flowchart.js, flowchart.min.js, and its dependency raphael.min.js
to the static/js folder.
Adds a plugin for jquery and flowchart.js to flowchart-jquery-plugin.js
to make using flow charts easier.
Moves the include of scripts.js to the bottom of the includes because
that's the only order that flowcharts would work with.
Adds new shortcode flowchart.html that creates a flowchart using
flowchart.js.