Replaces flowchart.js with mermaid.js
Deletes raphael.min.js, flowchart.js, and flowchart-jquery-plugin.js
from the site.
Adds mermaid.js and mermaid.min.js to the site.
Replaces the javascript includes at the bottom of the body of
baseof.html with the new include.
Updates the flowchart shortcode to use the mermaid class instead of the
flowchart-js class.
Adds javascript call to initialize mermaid to resume-material.js.