Adds basic TOC support in posts
Adds an adapted TOC partial that grabs header tags for use in single
post pages.
Adds show_toc param to post pages. If true, the TOC will be shown on the
page.
Sets show_toc to false in the starting-over post.
Adds adapted TOC partial as aside element that will only be shown if
show_toc is true.
Main content and TOC partial are now sharing a Materialize CSS Grid
together. The grid widths change based on the value of show_toc and the
size of the screen.